Identity column finder added to Phase I
Phase I is now more efficient because it is capable of finding existing identity matrix columns which reduces the number of artificial variables added and therefore the number of iterations needed. 

A mistake was found in which initial_solution was changing during Phase II due to the way Python stores arrays and lists in memory. This was also corrected by making a copy to allocate variables to different IDs.